Move readSession to sessions/core/
