+++ packages/core/src/sessions/index.ts
@@ -1,5 +1,5 @@
 export { createSession } from './core/create-session';
-export { readSession } from './read-session';
+export { readSession } from './core/read-session';
 export { Store, SessionStore, SessionAlreadyExists } from './session-store';
 export { Session } from './session';
 export { SessionState } from './session-state.interface';
