+++ packages/core/src/sessions/use-sessions.hook.spec.ts
@@ -29,7 +29,7 @@ import {
   SESSION_DEFAULT_INACTIVITY_TIMEOUT,
   SESSION_USER_COOKIE_NAME
 } from './constants';
-import { readSession } from './read-session';
+import { readSession } from './core/read-session';
 import { Session } from './session';
 import { SessionState } from './session-state.interface';
 import { SessionStore } from './session-store';
